New sounds, new sprites

+++ objects/oSellButton/Mouse_7.gml
@@ -1,5 +1,7 @@
-global.coins += global.sellprice;
-global.sellprice = 0;
-instance_destroy(global.instance_selected);
-deleteAllUpgradeBacks();
-global.instance_selected = noone;
+if (!global.gameover) {
+	global.coins += global.sellprice;
+	global.sellprice = 0;
+	instance_destroy(global.instance_selected);
+	deleteAllUpgradeBacks();
+	global.instance_selected = noone;
+}
