+++ objects/oTowerClickableParent/Mouse_10.gml
@@ -1,2 +1,4 @@
-global.new_tower_hover = object_index;
-deleteAllUpgradeBacks();
+if (!global.gameover) {
+	global.new_tower_hover = object_index;
+	deleteAllUpgradeBacks();
+}
